+++ src/statusbar.c
@@ -391,7 +391,7 @@ ncm_statusbar_main_hook(char *string, int32 string_len) {
 
 int32
 ncm_statusbar_message_delay_time(void) {
-    return (int32)Config.message_delay_time;
+    return Config.message_delay_time;
 }
 
 bool
